BREAKING CHANGE: Let errors bubble up rather than catching them
This should ensure the application using the client knows about any errors.
The application can then decide what to do with them, such as logging an
error or trying to recover.

This is a breaking change, as calls that would previously fail
without throwing errors now need to be handled with try/catch blocks.

Part of UserOfficeProject/stfc-user-office-project#231