performance: Avoid making duplicate requests
In many cases, the client gets several calls for the same function
with the same arguments in the space of a few microseconds.
For example, this happens when the frontend requests several GraphQL
resources which are reserved for User Officers. The backend will try to
check the user's token for each of those requests separately,
leading to many calls to the method for checking tokens with the same
argument.

This change allows the client to use a single UOWS call in these situations.
Each UOWS call is given a single ID, generated from the UOWS method name
and the arguments to the UOWS call. Calls for the same method, with the same
arguments, get the same ID. The client makes an async request the first time
it encounters a certain request ID, then caches the Promise until it receives
a response. Any request with the same ID will get the cached Promise, rather
than a new one generated from a new UOWS call. This avoids the need for
multiple identical requests while one is already in progress.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/580