Retry creating the UOWS client if creation failed
If the UOWS is down for any reason when we start the proposal backend,
the request used to generate the SOAP client will fail.

To recover from this problem, we retry creating the client
every 10 seconds if the last attempt failed. It does not matter
if the UOWS does down afterwards, as the first request is only used to
get the WSDL and does not create a permanent connection.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/580