performance: Make UOWS client a singleton class
This ensures that consumers can only create a single instance of the UOWS client.
The advantage of doing this is that the promise cache introduced in the previous
commits can be used more effectively, as all requests will go through a single
client and can be merged if they're identical. We also make ensure that the WSDL
is only fetched once and reduce memory usage, as there will not be
multiple clients each fetching, parsing and storing the WSDL of the service.
This may also make debugging easier, as there is a now single component
connecting to the external service rather than many.

The downside of this is that we can't have two UOWS clients for different URLs
in the same application. I can't think of a situation where we'd want to do that,
so I think the performance and ease of debugging are worth it.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/580